10 Close other multiselects on open
We now use the `:focus-within` selector to handle opening & closing the
dropdowns.
* `:focus-within` is supported by all major browsers as of January 2020
* The global `hideMultiselects` function has been removed since we
aren't using it. This is a breaking change, but it was never publicly
documented and I don't see much use case for it anymore. In case someone
was using `hideMultiselects`, they can now use
`document.activeElement.blur()` instead.